Moving defect stability to separate file
